general cell, alert view, shop tap and shop info

+++ Tokopedia/Classses/UI/CustomViews/ProgressBarView.h
@@ -10,6 +10,6 @@
 
 @interface ProgressBarView : UIView
 
-@property (nonatomic) NSInteger floatcount;
+@property (nonatomic) CGFloat floatcount;
 
 @end
